requirements.txt
absl-py==2.1.0
altgraph==0.17.4
annotated-types==0.7.0
astunparse==1.6.3
auto-py-to-exe==2.44.2
beautifulsoup4==4.12.3
blinker==1.9.0
blis==1.1.0
bottle==0.13.2
bottle-websocket==0.2.9
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.10
deep-translator==1.11.4
docutils==0.21.2
Eel==0.17.0
et_xmlfile==2.0.0
filelock==3.16.1
Flask==3.1.0
flatbuffers==24.12.23
fonttools==4.55.3
fsspec==2024.12.0
future==1.0.0
gast==0.6.0
gevent==24.10.3
gevent-websocket==0.10.1
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-language==2.16.0
google-pasta==0.2.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.1
grpcio-status==1.68.1
h5py==3.12.1
huggingface-hub==0.27.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
keras==3.7.0
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
Kivy-Garden==0.1.5
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
libclang==18.1.1
lxml==5.3.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
murmurhash==1.0.11
mysql-connector-python==9.2.0
namex==0.0.8
networkx==3.4.2
numpy==2.0.2
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.13.1
packaging==24.1
pandas==2.2.3
pefile==2023.2.7
pillow==11.0.0
pluggy==1.5.0
preshed==3.0.9
proto-plus==1.25.0
protobuf==5.29.2
psutil==6.1.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
pyparsing==3.2.0
pypiwin32==223
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pytest==8.3.5
pytest-html==4.1.1
pytest-metadata==3.1.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.2
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
razdel==0.5.0
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rsa==4.9
safetensors==0.4.5
setuptools==75.3.0
shellingham==1.5.4
six==1.16.0
smart-open==7.1.0
soupsieve==2.6
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.0
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tf_keras==2.18.0
thinc==8.3.3
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
transformers==4.47.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wheel==0.45.1
whichcraft==0.6.1
wrapt==1.17.0
XlsxWriter==3.2.0
zope.event==5.0
zope.interface==7.1.1
